Revert "Merge pull request #14 from HumphreyHCB/graalVersionAsOfNov2025"
This reverts commit 188a84cb7fd163adf450c7b4a848b393587bd9cf, reversing
changes made to fe8e5dde7119ab88addf3d9e664fc7d874a0a1b0.

+++ espresso/src/com.oracle.truffle.espresso.jdwp/src/com/oracle/truffle/espresso/jdwp/api/MethodRef.java
@@ -231,7 +231,7 @@ public interface MethodRef {
      *
      * @return last bci
      */
-    int getLastBCI();
+    long getLastBCI();
 
     /**
      * Determines if the method is a constructor.
